Fix early termination when last character of `recv()` is `]` or `}`
Can occur if the JSON response includes an array inside a map, or an array of maps, and the 4096th byte is the terminating character of that array or map. Now continues to read from the socket if the JSON decode fails.

May require a timeout in the future, since an invalid response might lead to an infinite loop.